Week 1/Arun/backend/app/pinecone_client.py
# backend/app/pinecone_client.py

import os
from pinecone import Pinecone, ServerlessSpec
from dotenv import load_dotenv

load_dotenv()

PINECONE_API_KEY = os.getenv("PINECONE_API_KEY")
PINECONE_INDEX_NAME = os.getenv("PINECONE_INDEX_NAME", "dental-appointments")

if not PINECONE_API_KEY:
    raise RuntimeError("ERROR: PINECONE_API_KEY is missing in environment variables.")

pc = Pinecone(api_key=PINECONE_API_KEY)

# Create index if it doesn't exist
if PINECONE_INDEX_NAME not in pc.list_indexes().names():
    pc.create_index(
        name=PINECONE_INDEX_NAME,
        dimension=64,              # MUST match DUMMY_VECTOR_DIM
        metric="cosine",
        spec=ServerlessSpec(
            cloud="aws",
            region="us-east-1"
        ),
    )

index = pc.Index(PINECONE_INDEX_NAME)
