Week 1/Arun/pyproject.toml
[project]
name = "arun"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic-ai",
    "pydantic[email]",
    "huggingface-hub",
    "nest-asyncio",
    "python-dotenv",
    "python-dateutil",
    "fastapi",
    "uvicorn[standard]",
    "npm>=0.1.1",
    "dateparser",
    "google-auth-oauthlib>=1.2.3",
    "google-auth-httplib2>=0.2.1",
    "google-api-python-client>=2.187.0",
]
